Modified the logging of various ATF messages
In ATF many messages were squelched to cut down on console
splat.  This was done by changing the logging level.  The
issue was that certain information that was of
interest was also squelched because it was at the same level
as the unwanted messages.  So code was modified to move the wanted
and unwanted messages to different logging levels so as to be able
to only display the wanted messages.

This allowed Jira item [RSOM-405] ManualTest> 20.01.01 – LPDDR4-3200 (
1.5 GHz) Support speed test failed. to be resolved

Signed-off-by: Richard Feliciano <rfeliciano@BeaconEmbedded.com>